+++ tests/hijack/email.js
@@ -1,25 +1,21 @@
-import { TestHelpers } from '../_helpers/helpers';
+import { addAsyncTest, callAsync, getLastMethodEvents, registerMethod } from '../_helpers/helpers';
 
-const Email = Package['email'].Email;
+async function sendTestEmailThroughMethod () {
+  const Email = Package['email'].Email;
 
-function sendTestEmailThroughMethod (func) {
-  const methodId = TestHelpers.registerMethod(function () {
-    return Email[func]({ from: 'arunoda@meteorhacks.com', to: 'hello@meteor.com' });
+  const methodId = registerMethod(async function () {
+    await Email.sendAsync({ from: 'arunoda@meteorhacks.com', to: 'hello@meteor.com' });
   });
 
-  const client = TestHelpers.getMeteorClient();
-
-  client.call(methodId);
+  await callAsync(methodId);
 }
 
-Tinytest.add(
+addAsyncTest(
   'Email - success',
-  function (test) {
-    TestHelpers.enableTrackingMethods();
+  async function (test) {
+    await sendTestEmailThroughMethod();
 
-    sendTestEmailThroughMethod('send');
-
-    const events = TestHelpers.getLastMethodEvents([0]);
+    const events = getLastMethodEvents([0]);
 
     const expected = [
       ['start'],
@@ -29,32 +25,5 @@ Tinytest.add(
     ];
 
     test.equal(events, expected);
-
-    TestHelpers.cleanTestData();
   }
 );
-
-if (Email.sendAsync) {
-  Tinytest.add(
-    'Email - sendAsync',
-    function (test) {
-      TestHelpers.enableTrackingMethods();
-
-      sendTestEmailThroughMethod('sendAsync');
-
-      const events = TestHelpers.getLastMethodEvents([0]);
-
-      const expected = [
-        ['start'],
-        ['wait'],
-        ['email'],
-        ['async'],
-        ['complete']
-      ];
-
-      test.equal(events, expected);
-
-      TestHelpers.cleanTestData();
-    }
-  );
-}