+++ tests/hijack/error.js
@@ -1,8 +1,14 @@
 import { Meteor } from 'meteor/meteor';
 import { ErrorModel } from '../../lib/models/errors';
-import { GetMeteorClient, RegisterMethod, RegisterPublication } from '../_helpers/helpers';
-
-const HTTP = Package['http'].HTTP;
+import {
+  addAsyncTest,
+  callAsync,
+  getMeteorClient,
+  registerMethod,
+  RegisterMethod,
+  registerPublication
+} from '../_helpers/helpers';
+const HTTP = Package['http']?.HTTP;
 
 Tinytest.add(
   'Errors - Meteor._debug - track with Meteor._debug',
@@ -62,17 +68,19 @@ Tinytest.add(
   }
 );
 
-Tinytest.add(
+addAsyncTest(
   'Errors - Meteor._debug - do not track method errors',
-  function (test) {
+  async function (test) {
     let originalErrorTrackingStatus = Kadira.options.enableErrorTracking;
+
     Kadira.enableErrorTracking();
+
     Kadira.models.error = new ErrorModel('foo');
+
     let method = RegisterMethod(causeError);
-    let client = GetMeteorClient();
 
     try {
-      client.call(method);
+      await callAsync(method);
     } catch (e) {
       // ignore the error
     }
@@ -96,8 +104,8 @@ Tinytest.addAsync(
     let originalErrorTrackingStatus = Kadira.options.enableErrorTracking;
     Kadira.enableErrorTracking();
     Kadira.models.error = new ErrorModel('foo');
-    let pubsub = RegisterPublication(causeError);
-    let client = GetMeteorClient();
+    let pubsub = registerPublication(causeError);
+    let client = getMeteorClient();
     client.subscribe(pubsub, {
       onError () {
         let payload = Kadira.models.error.buildPayload();
@@ -116,9 +124,9 @@ Tinytest.addAsync(
   }
 );
 
-Tinytest.addAsync(
+addAsyncTest(
   'Errors - Meteor._debug - do not track when no arguments',
-  function (test, done) {
+  async function (test) {
     let originalErrorTrackingStatus = Kadira.options.enableErrorTracking;
     Kadira.enableErrorTracking();
     Kadira.models.error = new ErrorModel('foo');
@@ -126,16 +134,15 @@ Tinytest.addAsync(
     let payload = Kadira.models.error.buildPayload();
     test.equal(0, payload.errors.length);
     _resetErrorTracking(originalErrorTrackingStatus);
-    done();
   }
 );
 
 // How Meteor gives errors to Meteor._debug versions 1.4 - 1.6
 // is already tested above. It changed in 1.7 which these tests covers
 if (!['1.4', '1.5', '1.6'].find(prefix => Meteor.release.startsWith(`METEOR@${prefix}`))) {
-  Tinytest.addAsync(
+  addAsyncTest(
     'Errors - Meteor._debug - preserve error thrown in Meteor.bindEnvironment',
-    function (test, done) {
+    async function (test) {
       let originalErrorTrackingStatus = Kadira.options.enableErrorTracking;
       Kadira.enableErrorTracking();
       const error = new Error('test');
@@ -147,7 +154,6 @@ if (!['1.4', '1.5', '1.6'].find(prefix => Meteor.release.startsWith(`METEOR@${pr
         test.equal(error.message, loggedError.message);
         test.equal(error.stack, loggedError.stack);
         _resetErrorTracking(originalErrorTrackingStatus);
-        done();
       };
 
       Meteor.bindEnvironment(function () {
@@ -156,18 +162,22 @@ if (!['1.4', '1.5', '1.6'].find(prefix => Meteor.release.startsWith(`METEOR@${pr
     }
   );
 
-  Tinytest.addAsync(
+  addAsyncTest(
     'Errors - Meteor._debug - track Meteor Error thrown in Meteor.bindEnvironment',
-    function (test, done) {
+    async function (test) {
       let originalErrorTrackingStatus = Kadira.options.enableErrorTracking;
       Kadira.enableErrorTracking();
       Kadira.models.error = new ErrorModel('foo');
       const error = new Meteor.Error('test');
+
       Meteor.bindEnvironment(function () {
         throw error;
       })();
+
       let payload = Kadira.models.error.buildPayload();
+
       let errorTrace = payload.errors[0];
+
       let expected = {
         appId: 'foo',
         name: 'Exception in callback of async function: [test]',
@@ -192,20 +202,23 @@ if (!['1.4', '1.5', '1.6'].find(prefix => Meteor.release.startsWith(`METEOR@${pr
 
       delete errorTrace.startTime;
       delete errorTrace.trace.at;
+
       test.equal(expected, errorTrace);
       _resetErrorTracking(originalErrorTrackingStatus);
-      done();
     }
   );
 }
 
-Tinytest.addAsync(
+addAsyncTest(
   'Errors - unhandledRejection - track unhandledRejection',
-  function (test, done) {
+  async function (test) {
     let originalErrorTrackingStatus = Kadira.options.enableErrorTracking;
+
     Kadira.enableErrorTracking();
     Kadira.models.error = new ErrorModel('foo');
+
     let error = new Error('rejected');
+
     Promise.reject(error);
 
     Meteor.defer(function () {
@@ -218,14 +231,13 @@ Tinytest.addAsync(
       test.equal(error.subType, 'unhandledRejection');
 
       _resetErrorTracking(originalErrorTrackingStatus);
-      done();
     });
   }
 );
 
-Tinytest.addAsync(
+addAsyncTest(
   'Errors - unhandledRejection - undefined reason',
-  function (test, done) {
+  async function (test) {
     let originalErrorTrackingStatus = Kadira.options.enableErrorTracking;
     Kadira.enableErrorTracking();
     Kadira.models.error = new ErrorModel('foo');
@@ -242,23 +254,24 @@ Tinytest.addAsync(
       test.equal(error.subType, 'unhandledRejection');
 
       _resetErrorTracking(originalErrorTrackingStatus);
-      done();
     });
   }
 );
 
-Tinytest.addAsync(
+addAsyncTest(
   'Errors - method error - track Meteor.Error',
-  function (test, done) {
+  async function (test) {
     let originalErrorTrackingStatus = Kadira.options.enableErrorTracking;
-    let methodId = RegisterMethod(function () {
+
+    let methodId = registerMethod(function () {
       throw new Meteor.Error('ERR_CODE', 'reason');
     });
-    let client = GetMeteorClient();
+
     try {
-      client.call(methodId);
+      await callAsync(methodId);
     } catch (ex) {
       let errorMessage = 'reason [ERR_CODE]';
+
       test.equal(ex.message, errorMessage);
       let payload = Kadira.models.error.buildPayload();
       let error = payload.errors[0];
@@ -267,23 +280,23 @@ Tinytest.addAsync(
       let lastEvent = error.trace.events[error.trace.events.length - 1];
       test.isTrue(lastEvent[2].error.message.indexOf(errorMessage) >= 0);
       test.isTrue(lastEvent[2].error.stack.indexOf(errorMessage) >= 0);
-      done();
     }
 
     _resetErrorTracking(originalErrorTrackingStatus);
   }
 );
 
-Tinytest.addAsync(
+addAsyncTest(
   'Errors - method error - store error details property',
-  function (test, done) {
+  async function (test) {
     let originalErrorTrackingStatus = Kadira.options.enableErrorTracking;
-    let methodId = RegisterMethod(function () {
+
+    let methodId = registerMethod(function () {
       throw new Meteor.Error('ERR_CODE', 'reason', 'details');
     });
-    let client = GetMeteorClient();
+
     try {
-      client.call(methodId);
+      await callAsync(methodId);
     } catch (ex) {
       let errorMessage = 'reason [ERR_CODE]';
       test.equal(ex.message, errorMessage);
@@ -296,23 +309,23 @@ Tinytest.addAsync(
       test.isTrue(lastEvent[2].error.message.indexOf(errorMessage) >= 0);
       test.isTrue(lastEvent[2].error.stack.indexOf(errorMessage) >= 0);
       test.equal(lastEvent[2].error.details, 'details');
-      done();
     }
 
     _resetErrorTracking(originalErrorTrackingStatus);
   }
 );
 
-Tinytest.addAsync(
+addAsyncTest(
   'Errors - method error - track NodeJs Error',
-  function (test, done) {
+  async function (test) {
     let originalErrorTrackingStatus = Kadira.options.enableErrorTracking;
+
     let methodId = RegisterMethod(function () {
       throw new Error('the-message');
     });
-    let client = GetMeteorClient();
+
     try {
-      client.call(methodId);
+      await callAsync(methodId);
     } catch (ex) {
       let errorMessage = 'the-message';
       test.isTrue(ex.message.match(/Internal server error/));
@@ -323,8 +336,99 @@ Tinytest.addAsync(
       let lastEvent = error.trace.events[error.trace.events.length - 1];
       test.isTrue(lastEvent[2].error.message.indexOf(errorMessage) >= 0);
       test.isTrue(lastEvent[2].error.stack.indexOf(errorMessage) >= 0);
-      done();
     }
+
+    _resetErrorTracking(originalErrorTrackingStatus);
+  }
+);
+
+addAsyncTest(
+  'Errors - method error - preserve async method promise',
+  async function (test) {
+    let methodPromise;
+
+    let methodId = RegisterMethod(function () {
+      methodPromise = Promise.resolve('result');
+      return methodPromise;
+    });
+
+    let returnedPromise = Meteor.server.method_handlers[methodId]();
+
+    test.isTrue(
+      returnedPromise === methodPromise,
+      'the method wrapper should return the original promise'
+    );
+    test.equal(await returnedPromise, 'result');
+  }
+);
+
+addAsyncTest(
+  'Errors - method error - track NodeJs Error thrown in async method',
+  async function (test) {
+    let originalErrorTrackingStatus = Kadira.options.enableErrorTracking;
+
+    Kadira.enableErrorTracking();
+    Kadira.models.error = new ErrorModel('foo');
+
+    let methodId = RegisterMethod(async function () {
+      await Promise.resolve();
+      throw new Error('async-the-message');
+    });
+
+    let hadError = false;
+
+    try {
+      await callAsync(methodId);
+    } catch (ex) {
+      hadError = true;
+      test.isTrue(ex.message.match(/Internal server error/), `client should see the sanitized error: ${ex.message}`);
+    }
+
+    test.isTrue(hadError, 'the method should have thrown');
+
+    let payload = Kadira.models.error.buildPayload();
+
+    let trackedErrors = payload.errors.map(e => [e.type, e.subType, e.name]);
+    test.equal(payload.errors.length, 1, `the error should be tracked exactly once: ${JSON.stringify(trackedErrors)}`);
+
+    let error = payload.errors[0];
+    test.equal(error.type, 'method');
+    test.equal(error.subType, methodId);
+    test.isTrue(error.name.indexOf('async-the-message') >= 0, `error name should have the real message: ${error.name}`);
+    test.isTrue(error.stacks[0].stack.indexOf('async-the-message') >= 0, 'error stack should be the real stack');
+
+    let lastEvent = error.trace.events[error.trace.events.length - 1];
+    test.isTrue(
+      lastEvent[2].error.message.indexOf('async-the-message') >= 0,
+      `trace error event should have the real message: ${JSON.stringify(lastEvent[2])}`
+    );
+
+    _resetErrorTracking(originalErrorTrackingStatus);
+  }
+);
+
+addAsyncTest(
+  'Errors - unhandledRejection - null reason',
+  async function (test) {
+    let originalErrorTrackingStatus = Kadira.options.enableErrorTracking;
+
+    Kadira.enableErrorTracking();
+    Kadira.models.error = new ErrorModel('foo');
+
+    let threw = false;
+
+    try {
+      process.emit('unhandledRejection', null, Promise.resolve());
+    } catch (ex) {
+      threw = true;
+    }
+
+    test.equal(threw, false);
+
+    let payload = Kadira.models.error.buildPayload();
+
+    test.equal(payload.errors.length, 1);
+    test.equal(payload.errors[0] && payload.errors[0].subType, 'unhandledRejection');
 
     _resetErrorTracking(originalErrorTrackingStatus);
   }