+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Next
+* Support Meteor 3
+* Add `Monti.event`
+* Tracer supports parallel events
+* Improve support for proxies
+
 ## v2.50.2
 
 - Much better support for apps that have a very large number of traces or have traces that are very large
@@ -14,7 +20,6 @@ February 3, 2025
 * Metrics for bullmq jobs will now be grouped by the combination of the queue name and job name
 
 ## v2.50.0
-
 December 30, 2024
 
 * Create custom traces with `Monti.traceJob`